+++ extensions/markdown-language-features/package.json
@@ -8,7 +8,7 @@
   "license": "MIT",
   "aiKey": "0c6ae279ed8443289764825290e4f9e2-1a736e7c-1324-4338-be46-fc2a58ae4d14-7255",
   "engines": {
-    "vscode": "^1.20.0"
+    "vscode": "^1.70.0"
   },
   "main": "./out/extension",
   "browser": "./dist/browser/extension",
